+++ .evergreen/run-big-endian-test.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-source $DRIVERS_TOOLS/.evergreen/init-node-and-npm-env.sh
-
-npx mocha test/s390x/big_endian.test.ts